email_automation_mvp/requirements.txt
requests
beautifulsoup4
pandas
lxml
playwright
pdfminer.six
python-docx
python-pptx
